// whirlpool/hotpointuk/Custom-apps/BE/worldpayconnector/node/typings/configs.d.ts
export interface configs {
  vtexAppKey: string,
  vtexAppToken: string,
  wpUrl: string,
  merchantCode: string,
  installationID: string,
  wpusername: string,
  wppassword: string,
  paymentAppName: string,
  locale: string,
  wpVersion: string,
  gcpappkey: string,
  gcpapptoken: string,
  useXipay: boolean
  xipayUrl: string,
  xipayUsername: string,
  xipayPassword: string,
  xipayMerchantid: string,
  sapmid: string,
  sapgl: string
}